+++ R/RxODE.R
@@ -4,6 +4,10 @@ R_NegInf <- -Inf # nolint
 R_PosInf <- Inf # nolint
 
 .linCmtSens <- NULL
+.clearME <- function(){
+  assignInMyNamespace(".rxMECode", "")
+}
+
 ##' Create an ODE-based model specification
 ##'
 ##' Create a dynamic ODE-based model object suitably for translation
@@ -331,6 +335,7 @@ R_PosInf <- Inf # nolint
 ##' @importFrom methods signature is
 ##' @importFrom memoise memoise
 ##' @importFrom utils capture.output
+##' @importFrom qs qsave
 ##' @import tools
 ##' @export
 RxODE <- # nolint
@@ -420,11 +425,7 @@ RxODE <- # nolint
       }
     }
     .env <- new.env(parent = baseenv())
-    .env$.rxTransCode <- "SEXP matLst = PROTECT(allocVector(VECSXP, 0));pro++;\n SET_VECTOR_ELT(lst,  17, matLst);\n"
-    .setTransCode(.env$.rxTransCode, "")
-    on.exit(.setTransCode("bad", ""), add = TRUE)
     .env$.mv <- rxGetModel(model, calcSens = calcSens, calcJac = calcJac, collapseModel = collapseModel, indLin = indLin)
-    .env$.rxTransCode <- .rxTransCode
     assignInMyNamespace(".linCmtSens", linCmtSens)
     if (.Call(`_RxODE_isLinCmt`) == 1L) {
       .env$.linCmtM <- rxNorm(.env$.mv)
@@ -484,10 +485,9 @@ RxODE <- # nolint
         if (file.exists(wd)) {
           setwd(wd)
         }
-        .rx$.setTransCode(get(".rxTransCode", envir = .(.env)))
         on.exit({
           setwd(.lwd)
-          .rx$.setTransCode("bad")
+          .rx$.clearME()
         })
         .rx$.extraC(extraC)
         if (missing.modName) {
@@ -871,10 +871,9 @@ rxGetModel <- function(model, calcSens = NULL, calcJac = NULL, collapseModel = N
     }
   }
   if (indLin) {
-    .ret0 <- .rxIndLin(.ret)
-    .new <- paste0(rxNorm(.ret), "\n", .ret0[[3]])
-    .setTransCode(.ret0[[4]], .ret0[[3]])
-    .Call(`_RxODE_clearTrans`)
+    .code <- .rxIndLin(.ret)
+    .new <- paste0(rxNorm(.ret), "\n", .code)
+    assignInMyNamespace(".rxMECode", .code)
     .ret <- rxModelVars(.new)
   }
   return(.ret)
@@ -1121,7 +1120,7 @@ rxMd5 <- function(model, # Model File
     .ret <- c(.ret, RxODE::rxVersion())
     return(list(
       text = model,
-      digest = digest::digest(.ret, serialize = TRUE, algo = "md5")
+      digest = digest::digest(list(.ret, .indLinInfo), serialize = TRUE, algo = "md5")
     ))
   } else {
     RxODE::rxModelVars(model)$md5
@@ -1194,19 +1193,7 @@ rxTrans.default <- function(model,
   }
 }
 
-.rxTransCode <- "bad"
 .rxMECode <- ""
-##' Set translation extra code
-##'
-##' @param x Code to set
-##' @return nothing
-##' @author Matthew Fidler
-##' @keywords internal
-##' @noRd
-.setTransCode <- function(x = "bad", y = "") {
-  assignInMyNamespace(".rxTransCode", x)
-  assignInMyNamespace(".rxMECode", y)
-}
 
 ##' @rdname rxTrans
 ##' @export
@@ -1229,7 +1216,7 @@ rxTrans.character <- memoise::memoise(function(model,
   .ret <- .Call(
     `_RxODE_trans`, model, modelPrefix, md5, .isStr,
     as.integer(crayon::has_color()),
-    .rxTransCode, .rxMECode, .rxSupportedFuns()
+    .rxMECode, .rxSupportedFuns()
   )
   if (inherits(.ret, "try-error")) {
     message("model")
@@ -1482,18 +1469,23 @@ rxCompile.rxModelVars <- function(model, # Model
         if (!is.null(modName)) {
           .newMod <- regexpr("_new", modName) != -1
         }
+        .rxModelVarsLast[[18]] <- .indLinInfo
+        .model <- .rxModelVarsLast$model
+        .model["indLin"] <- .rxMECode
+        .rxModelVarsLast$model <- .model
         if (!is.null(package) & !.newMod) {
           .libname <- c(package, gsub(.Platform$dynlib.ext, "", basename(.cDllFile)))
           .Call(
             `_RxODE_codegen`, .cFile, prefix, .libname,
-            .trans["parsed_md5"], paste(.rxTimeId(.trans["parsed_md5"])))
+            .trans["parsed_md5"], paste(.rxTimeId(.trans["parsed_md5"])),
+            .rxModelVarsLast)
         } else {
           .libname <- gsub(.Platform$dynlib.ext, "", basename(.cDllFile))
           .libname <- c(.libname, .libname)
           .Call(
             `_RxODE_codegen`, .cFile, prefix, .libname,
-            .trans["parsed_md5"], paste(.rxTimeId(.trans["parsed_md5"]))
-          )
+            .trans["parsed_md5"], paste(.rxTimeId(.trans["parsed_md5"])),
+            .rxModelVarsLast)
         }
         .defs <- ""
         .ret <- sprintf(
@@ -1734,6 +1726,7 @@ rxNorm <- function(obj, condition = NULL, removeInis, removeJac, removeSens) {
 
 
 .rxModelVarsCCache <- NULL
+.rxModelVarsLast <- NULL
 .rxModelVarsCharacter <- function(obj) {
   if (length(obj) == 1) {
     .parseModel <- tempfile("parseModel4")
@@ -1752,6 +1745,7 @@ rxNorm <- function(obj, condition = NULL, removeInis, removeJac, removeSens) {
     .ret <- rxTrans(.parseModel, modelPrefix = .prefix, modVars = TRUE)
     .cFile <- list(.exists, ifelse(.exists, obj, ""), .prefix)
     assignInMyNamespace(".rxModelVarsCCache", .cFile)
+    assignInMyNamespace(".rxModelVarsLast", .ret)
     return(.ret)
   } else {
     .rxModelVarsCharacter(paste(obj, collapse = "\n"))