+++ src/RcppExports.cpp
@@ -521,6 +521,28 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// rxQs
+Rcpp::CharacterVector rxQs(SEXP const x);
+RcppExport SEXP _RxODE_rxQs(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP const >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(rxQs(x));
+    return rcpp_result_gen;
+END_RCPP
+}
+// rxQr
+SEXP rxQr(const std::string& encoded_string);
+RcppExport SEXP _RxODE_rxQr(SEXP encoded_stringSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const std::string& >::type encoded_string(encoded_stringSEXP);
+    rcpp_result_gen = Rcpp::wrap(rxQr(encoded_string));
+    return rcpp_result_gen;
+END_RCPP
+}
 // rxIs
 bool rxIs(const RObject& obj, std::string cls);
 static SEXP _RxODE_rxIs_try(SEXP objSEXP, SEXP clsSEXP) {