+++ src/init.c
@@ -37,10 +37,9 @@ SEXP _rxTick();
 SEXP _rxProgressStop(SEXP);
 SEXP _rxProgressAbort(SEXP);
 SEXP _RxODE_codeLoaded();
-SEXP _RxODE_clearTrans();
 
 SEXP _RxODE_trans(SEXP parse_file, SEXP prefix, SEXP model_md5, SEXP isStr, SEXP, SEXP, SEXP, SEXP);
-SEXP _RxODE_codegen(SEXP c_file, SEXP prefix, SEXP libname, SEXP pMd5, SEXP timeId);
+SEXP _RxODE_codegen(SEXP c_file, SEXP prefix, SEXP libname, SEXP pMd5, SEXP timeId, SEXP lastMv);
 SEXP _RxODE_parseModel(SEXP type);
 SEXP _RxODE_isLinCmt();
 SEXP _RxODE_RcppExport_registerCCallable();
@@ -203,6 +202,10 @@ SEXP _RxODE_convertId_(SEXP);
 SEXP _RxODE_rpp_(SEXP nS, SEXP lambdaS, SEXP gammaS, SEXP probS, SEXP t0S,
 		 SEXP tmaxS, SEXP randomOrderS);
 
+SEXP _RxODE_rxQs(SEXP);
+
+SEXP _RxODE_rxQr(SEXP);
+
 extern int rxIsCurrentC(SEXP obj);
 
 rx_solve *getRxSolve_();
@@ -330,10 +333,9 @@ void R_init_RxODE(DllInfo *info){
     {"_rxTick", (DL_FUNC) &_rxTick, 0},
     {"_rxProgressStop", (DL_FUNC) &_rxProgressStop, 1},
     {"_rxProgressAbort", (DL_FUNC) &_rxProgressAbort, 1},
-    {"_RxODE_trans", (DL_FUNC) &_RxODE_trans, 8},
-    {"_RxODE_codegen", (DL_FUNC) &_RxODE_codegen, 5},
+    {"_RxODE_trans", (DL_FUNC) &_RxODE_trans, 7},
+    {"_RxODE_codegen", (DL_FUNC) &_RxODE_codegen, 6},
     {"_RxODE_codeLoaded", (DL_FUNC) &_RxODE_codeLoaded, 0},
-    {"_RxODE_clearTrans", (DL_FUNC) &_RxODE_clearTrans, 0},
     {"_RxODE_parseModel", (DL_FUNC) &_RxODE_parseModel, 1},
     {"_RxODE_isLinCmt", (DL_FUNC) &_RxODE_isLinCmt, 0},
     {"RxODE_get_mv", (DL_FUNC) &RxODE_get_mv, 0},
@@ -459,9 +461,12 @@ void R_init_RxODE(DllInfo *info){
     {"_probit", (DL_FUNC) _probit, 3},
     {"_probitInv", (DL_FUNC) _probitInv, 3},
     {"_RxODE_rxrandnV", (DL_FUNC) _RxODE_rxrandnV, 2},
+    {"_RxODE_rxQs", (DL_FUNC) _RxODE_rxQs, 1},
+    {"_RxODE_rxQr", (DL_FUNC) _RxODE_rxQr, 1},
     {NULL, NULL, 0}
   };
   // C callable to assign environments.
+  R_RegisterCCallable("RxODE","_RxODE_rxQr", (DL_FUNC) &_RxODE_rxQr);
   R_RegisterCCallable("RxODE", "simeps", (DL_FUNC) &simeps);
   R_RegisterCCallable("RxODE", "simeta", (DL_FUNC) &simeta);
   R_RegisterCCallable("RxODE", "getSilentErr", (DL_FUNC) &getSilentErr);